Separating printing between base and zuul.
Getting rid of the raw printers too. They do not provide much and are something to maintain.
